% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/some_functions.R
\name{this_function}
\alias{this_function}
\title{A function that does nothing}
\usage{
this_function(x)
}
\arguments{
\item{x}{Anything you want}
}
\value{
\code{x} invisibly
}
\description{
this_function exists to demonstrate documentation generation with roxygen.
See \url{} for more details.
}
\section{Contributors}{

All these people know how to contribute to the docs for \code{this_function}:
}

\examples{
this_function(1)
}
